% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SG.R
\name{calcgrad}
\alias{calcgrad}
\title{Calculate gradient between two points on stream}
\usage{
calcgrad(points, ras, fctr)
}
\arguments{
\item{points}{Data Frame, output from getupdown function or same format as output from getupdown function}

\item{ras}{Raster of DEM. Works best if it is in the same projection as original points and lines
shapefiles used to calculate the stationing.}

\item{fctr}{Factor to multiply the segment length by. For example, if lengths are in meters and you want
gradient in meters/km fctr must be 0.001}
}
\description{
Calculate gradient between two points on stream
}
\details{
Calculates the stream gradient of segments on a line delimited by an upstream and downstream point. This
function works best if the points input is directly used from the output of the getupdown function.
}
