% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SG.R
\name{calcpoint}
\alias{calcpoint}
\title{Calculates the x and y coordinates of a point along a line}
\usage{
calcpoint(points, distp)
}
\arguments{
\item{points}{A 2x4 matrix where the first two columns specify the
x and y values for the endpoints of a line, the third column is the,
segment distance for the segment where the point is the ending segment,
and  the fourth column specifies the distance that vertex is along the
whole line (0 if it is the beginning of a line)}

\item{distp}{The distance along the entire line of the point to be
calculated}
}
\description{
Calculates the x and y coordinates of a point along a line
}
\details{
This calculates the distance along a line using the output from
the getdists() function using vector algebra. See
https://math.stackexchange.com/questions/175896/finding-a-point-along-a-line-a-certain-distance-away-from-another-point
for a mathematical explanation
}
