% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SG.R
\name{calculatestationing}
\alias{calculatestationing}
\title{Calculate stationing (distance) along line}
\usage{
calculatestationing(points, lines, maxDist = NA)
}
\arguments{
\item{points}{A SpatialPointsDataFrame in a planar (i.e. UTM) projection}

\item{lines}{A SpatialLinesDataFrame in a planar (i.e. UTM) projection}

\item{maxDist}{Maximum distance to be input into snapPointsToLines function}
}
\description{
Calculate stationing (distance) along line
}
\details{
See http://rstudio-pubs-static.s3.amazonaws.com/12524_7de6eb887f2945389c5d12869388be14.html
for original code and explanation
}
