% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SG.R
\name{getdists}
\alias{getdists}
\title{Calculate segment lengths and total line length}
\usage{
getdists(coords)
}
\arguments{
\item{coords}{A matrix of x (first column) and y (second column) coordinates
of vertices along a line.

Calculates the segment length of each segment along a multipart line from the
coordinates of the vertices of that line using the equation
sqrt((x2 - x1)^2 + sqrt(y2 - y1)^2)}
}
\description{
Calculate segment lengths and total line length
}
