% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SG.R
\name{getupdown}
\alias{getupdown}
\title{Calculate the x and y values of points a specified distance
upstream and downstream a point on a line}
\usage{
getupdown(points, lines, seglength, idcol = NULL)
}
\arguments{
\item{points}{A SpatialPointsDataFrame in a planar projection}

\item{lines}{A SpatialLinesDataFrame in a planar projection}

\item{seglength}{Numeric 1L The distance upstream and downstream to calculate new points}

\item{idcol}{Character, optional column name for an identification column from points input}
}
\description{
Calculate the x and y values of points a specified distance
upstream and downstream a point on a line
}
\details{
This function calculates the x and y values that are d units away from points a known
distance along a lines spatial object. NOTE: it will only work if the input spatial data are in
a planar projection (i.e. UTM).
}
