% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SG.R
\name{pointonline}
\alias{pointonline}
\title{See if point is on Line}
\usage{
pointonline(point, line_start, line_end)
}
\arguments{
\item{point}{pair of coordinates (x,y)}

\item{line_start}{pair of coordinates (x,y)}

\item{line_end}{pair of coordinates (x,y)}
}
\description{
See if point is on Line
}
\details{
See http://rstudio-pubs-static.s3.amazonaws.com/12524_7de6eb887f2945389c5d12869388be14.html
for explanation and original code
}
